Fix(all): 모든 부분 연관된 버그 수정

+++ src/main/java/com/nbc/trello/comment/CommentController.java
@@ -1,9 +1,9 @@
 package com.nbc.trello.comment;
 
+import com.nbc.trello.board.service.BoardService;
 import com.nbc.trello.card.entity.Card;
-import com.nbc.trello.global.dto.ApiResponse;
+import com.nbc.trello.global.response.ApiResponse;
 import com.nbc.trello.users.UserDetailsImpl;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,7 +11,6 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -19,35 +18,59 @@ import java.util.List;
 @RequestMapping("/boards/{boardId}/columns/{columnId}/cards/{cardId}/comments")
 public class CommentController {
     private final CommentService commentService;
+    private final BoardService boardService;
 
     @PostMapping
-    public ResponseEntity<ApiResponse<Void>> createComment(@AuthenticationPrincipal UserDetailsImpl userDetails, Card card, CommentRequestDto commentRequestDto){
+    public ResponseEntity<ApiResponse<Void>> createComment(
+            @AuthenticationPrincipal UserDetailsImpl userDetails,
+            Card card,
+            CommentRequestDto commentRequestDto,
+            @PathVariable Long boardId
+    ) {
+        boardService.checkAuthorization(userDetails.getUser(), boardId);
+
         commentService.createComment(userDetails, card, commentRequestDto);
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(ApiResponse.of(HttpStatus.CREATED.value(), "success", null ));
+                .body(ApiResponse.of(HttpStatus.CREATED.value(), "success", null));
     }
 
     @PatchMapping("/{commentId}")
-    public ResponseEntity<ApiResponse<Void>> updateComment(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable Long commentId, @RequestBody CommentRequestDto commentRequestDto) throws Exception {
-        commentService.updateComment(userDetails,commentId,commentRequestDto);
+    public ResponseEntity<ApiResponse<Void>> updateComment(
+            @AuthenticationPrincipal UserDetailsImpl userDetails,
+            @PathVariable Long commentId,
+            @RequestBody CommentRequestDto commentRequestDto,
+            @PathVariable Long boardId
+    ) throws Exception {
+        boardService.checkAuthorization(userDetails.getUser(), boardId);
+        commentService.updateComment(userDetails, commentId, commentRequestDto);
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(ApiResponse.of(HttpStatus.OK.value(), "success", null ));
+                .body(ApiResponse.of(HttpStatus.OK.value(), "success", null));
     }
 
     @DeleteMapping("/{commentId}")
-    public ResponseEntity<ApiResponse<Void>> deleteComment(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable Long commentId) throws Exception {
-        commentService.deleteComment(userDetails,commentId);
+    public ResponseEntity<ApiResponse<Void>> deleteComment(
+            @AuthenticationPrincipal UserDetailsImpl userDetails,
+            @PathVariable Long commentId,
+            @PathVariable Long boardId
+    ) throws Exception {
+        boardService.checkAuthorization(userDetails.getUser(), boardId);
+        commentService.deleteComment(userDetails, commentId);
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(ApiResponse.of(HttpStatus.OK.value(), "success", null ));
+                .body(ApiResponse.of(HttpStatus.OK.value(), "success", null));
     }
 
     @GetMapping
-    public ResponseEntity<ApiResponse<List<CommentResponseDto>>> getComment(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable Long cardId) throws Exception {
+    public ResponseEntity<ApiResponse<List<CommentResponseDto>>> getComment(
+            @AuthenticationPrincipal UserDetailsImpl userDetails,
+            @PathVariable Long cardId,
+            @PathVariable Long boardId
+    ) throws Exception {
+        boardService.checkAuthorization(userDetails.getUser(), boardId);
         List<CommentResponseDto> commentResponseDto = commentService.getComment(cardId);
